+++ gd-gltg/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Improved the components and config relations between `gd-gltg` and `gd-geostreaming`
   [GEOD-1341](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1341)
-
+- Get version number in footer from package.json
+  [GLGVO-666](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-666)
 ## [3.3.0] - 2020-06-18
 
 ### Added